requirements.txt
flask>=1.0.4
importlib-metadata
nest-asyncio
plotly >=5.0.0
requests
retrying
setuptools
typing-extensions >=4.1.1

dash-html-components >=2.0.0
dash_table >=5.0.0
dash-core-components >=2.0.0

dash>=2
dash[ci,dev,testing]>=2.0
pyyaml>=5.0
pytest<8.1.0
wheel
dash>=2.9.0
gunicorn>=19.9.0
pandas>=1.2.3
pyyaml>=3.13
dash-bootstrap-components>=1.0.0
dash-mantine-components
dash-iconify
colorlover
statsmodels
python-frontmatter>=1.1.0
markdown2dash
pydantic
gunicorn>=21.2.0

